requirements.txt
keyboard==0.13.5
PyYAML==6.0.1
scapy==2.5.0
tabulate==0.9.0
